fix : 오류 디버깅

+++ ad-hell/src/main/java/com/adhd/ad_hell/domain/boardComment/command/domain/aggregate/BoardComment.java
@@ -22,7 +22,7 @@ public class BoardComment extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
-    private BigInteger Id;
+    private Long Id;    // ← BigDecimal/BigInteger → Long 으로 변경
 
     @Column(name = "comment_content", nullable = false)
     private String content;
